requirements.txt
Django==5.1.3
django-cleanup==9.0.0
django-debug-toolbar==4.4.6
python-dotenv==1.0.1
sorl-thumbnail==12.11.0
